test: fix broken tests, clean up game service methods
