+++ app/Models/Publisher.php
@@ -2,10 +2,12 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Publisher extends Model
 {
+    use HasFactory;
     /**
      * Get the Publisher
      */
