+++ app/Services/GameService.php
@@ -2,7 +2,6 @@
 
 namespace App\Services;
 use App\Models\Game;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Mockery;
 use Mockery\MockInterface;
 
@@ -14,31 +13,18 @@ class GameService {
 
     function getGamesById (int $id)
     {
-        $game = Game::findOrFail($id);
-        return $game;
-
+        return Game::findOrFail($id);
     }
 
     // todo: add optional params for orderBy, direction, and limit
     function getGames()
     {
-        $games = Game::all()->sortBy('title');
-
-        echo $games;
-        // $games = Game::factory()->make();
-        return $games;
-
+        return Game::all()->sortBy('title');
     }
 
     function searchGamesByTitle (string $searchTitle)
     {
-        $titleName = null;
-        $titleName = Game::where('title','LIKE',$searchTitle)->get();
-
-
-        return $titleName;
-
-
+        return Game::where('title','LIKE',$searchTitle)->get();
     }
 
     public function test_mock_gameservice()
